requirements.txt
langchain-community
langchain-openai
chromadb
pypdf
pandas
streamlit
python-dotenv